new: Minimal Expression.Indexed pipeline. Better lang type conversion

+++ src/nonim/backend/test/cases/control_if/expected.c
@@ -1,4 +1,4 @@
-static int abs_val (int const x) {
+static int64_t abs_val (int64_t const x) {
   if (x < 0) {
     return -x;
   } else {
