+++ src/nonim/backend/test/cases/control_while/expected.c
@@ -1,5 +1,5 @@
-static void countdown (int const n) {
-  int current = n;
+static void countdown (int64_t const n) {
+  int64_t current = n;
   while (0 < current) {
     current = current - 1;
   }
