+++ src/nonim/backend/test/cases/expression_call/expected.c
@@ -1,6 +1,6 @@
-static int add (int const x, int const y) {
+static int64_t add (int64_t const x, int64_t const y) {
   return x + y;
 }
-static int main () {
+static int64_t main () {
   return add(1, 2);
 }
