+++ src/nonim/backend/test/cases/expression_call/expected.zig
@@ -1,6 +1,6 @@
-fn add (x: i64, y: i64) i64 {
+fn add (x: isize, y: isize) isize {
   return x + y;
 }
-fn main () i64 {
+fn main () isize {
   return add(1, 2);
 }
