+++ src/nonim/backend/test/cases/procedure/expected.c
@@ -1,1 +1,1 @@
-static int add (int const x, int const y);
+static int64_t add (int64_t const x, int64_t const y);
