+++ src/nonim/backend/test/cases/procedure/expected.zig
@@ -1,1 +1,1 @@
-fn add (x: i64, y: i64) i64;
+fn add (x: isize, y: isize) isize;
