+++ src/nonim/backend/test/cases/statement_continue/expected.c
@@ -1,5 +1,5 @@
-static void skip_negatives (int const n) {
-  int current = 0;
+static void skip_negatives (int64_t const n) {
+  int64_t current = 0;
   while (current < n) {
     current = current + 1;
     if (current < 0) {
