+++ src/nonim/backend/test/cases/statement_continue/expected.zig
@@ -1,5 +1,5 @@
-fn skip_negatives (n: i64) void {
-  var current: i64 = 0;
+fn skip_negatives (n: isize) void {
+  var current: isize = 0;
   while (current < n) {
     current = current + 1;
     if (current < 0) {
